+++ include/TenshiEngine/Components.hpp
@@ -8,6 +8,6 @@
 #include "TenshiEngine/Engine/Components/Velocity.hpp"
 #include "TenshiEngine/Engine/Components/Grounded.hpp"
 #include "TenshiEngine/Engine/Components/Animation.hpp"
-#include "TenshiEngine/Engine/Components/PlayerState.hpp"
+#include "TenshiEngine/Engine/Components/State.hpp"
 #include "TenshiEngine/Engine/Components/HorizontalMovement.hpp"
 #include "TenshiEngine/Engine/Components/Jump.hpp"