+++ include/TenshiEngine/Engine/Systems/MovementSystem.hpp
@@ -2,7 +2,6 @@
 
 #include "System.hpp"
 #include "TenshiEngine/Engine/Core/Registry.hpp"
-#include <SFML/System/Vector2.hpp>
 
 namespace te{
 namespace systems{
@@ -13,7 +12,7 @@ namespace systems{
  * Checks playerInput, horizonzalMovement, jump, grounded, velocity 
  * and then calculates new velocity 
  */
-class MovementSystem : public System{
+class MovementSystem : public FixedSystem{
 public:
     /**
      * @brief Constructor
@@ -21,17 +20,19 @@ public:
      * creates the System
      *
      * @param registry reference to the registry
+     * @param deltaTime Fixed deltaTime
      */
-    MovementSystem(Registry &registry);
+    MovementSystem(Registry &registry, const float deltaTime);
 
-    void update() override; 
+    void update(); 
 
     ///Name of system
     static const std::string name;
 
 private:
     void updateVelocity(EntityID entityID);
-    float updateHorizontalMovement(components::PlayerInput *playerInput, components::HorizontalMovement *horizontalMovement);
+    float updateHorizontalMovement(EntityID entityID, components::PlayerInput *playerInput, components::HorizontalMovement *horizontalMovement);
+    void updateState(EntityID entityID, components::Velocity *velocity);
     float updateJump(EntityID entityID, components::PlayerInput *playerInput, components::Velocity *velocity);
 };