+++ include/TenshiEngine/Engine/Systems/PhysicsSystem.hpp
@@ -12,7 +12,7 @@ namespace systems{
  * Checks velocity, rigidBody, collider and then calculates
  * new velocity and new sprite position
  */
-class PhysicsSystem : public System{
+class PhysicsSystem : public FixedSystem{
 public:
     /**
      * @brief Constructor
@@ -20,8 +20,9 @@ public:
      * creates the System with defaultGravity(10.f)
      *
      * @param registry reference to the registry
+     * @param deltaTime Fixed deltaTime
      */
-    PhysicsSystem(Registry &registry);
+    PhysicsSystem(Registry &registry, const float deltaTime);
 
     /**
      * @brief Constructor
@@ -29,23 +30,24 @@ public:
      * creates the System with given gravity
      *
      * @param registry reference to the registry
+     * @param deltaTime deltaTime
      * @param gravity value of gravity
      */
-    PhysicsSystem(Registry &registry, float gravity);
+    PhysicsSystem(Registry &registry, float deltaTime, float gravity);
 
-    void update() override; 
+    void update(); 
 
     ///Name of system
     static const std::string name;
 
 private:
     void updateGravity(EntityID entityID, components::Velocity *velocity);
-    void updateCollision(EntityID entityID1, EntityID entityID2);
+    void checkCollision(EntityID entityID1, EntityID entityID2);
     void correctPosition(EntityID entityID, components::Transform *transform1, components::Transform *transform2, sf::FloatRect overlapRect);
 
     float mGravity;
 
-    inline static const float defaultGravity = 10.f;
+    inline static const float defaultGravity = 25.f;
 };
 
 }