+++ include/TenshiEngine/Systems.hpp
@@ -2,7 +2,6 @@
 
 #include "TenshiEngine/Engine/Systems/System.hpp"
 #include "TenshiEngine/Engine/Systems/AnimationSystem.hpp"
-#include "TenshiEngine/Engine/Systems/DebugRenderSystem.hpp"
 #include "TenshiEngine/Engine/Systems/RenderSystem.hpp"
 #include "TenshiEngine/Engine/Systems/InputSystem.hpp"
 #include "TenshiEngine/Engine/Systems/MovementSystem.hpp"