+++ src/Engine/Systems/MovementSystem.cpp
@@ -3,7 +3,7 @@
 namespace te{
 namespace systems{
 
-MovementSystem::MovementSystem(Registry &registry) : System(registry){}
+MovementSystem::MovementSystem(Registry &registry, float deltaTime) : FixedSystem(registry, deltaTime){}
 
 void MovementSystem::update(){
     const auto &entityIDs = mRegistry.view<components::Velocity>();
@@ -14,7 +14,7 @@ void MovementSystem::update(){
 
 void MovementSystem::updateVelocity(EntityID entityID){
     auto velocity = mRegistry.getComponent<components::Velocity>(entityID);
-    sf::Vector2f newVelocity = velocity->value;
+    auto oldValue = velocity->value;
 
     if(mRegistry.hasComponent<components::HorizontalMovement>(entityID)){
         auto horizontalMovement = mRegistry.getComponent<components::HorizontalMovement>(entityID);
@@ -22,44 +22,77 @@ void MovementSystem::updateVelocity(EntityID entityID){
         if(mRegistry.hasComponent<components::PlayerInput>(entityID)){
             auto playerInput = mRegistry.getComponent<components::PlayerInput>(entityID);
 
-            newVelocity.x = updateHorizontalMovement(playerInput, horizontalMovement);
-            
+            velocity->value.x = updateHorizontalMovement(entityID, playerInput, horizontalMovement);
             if(mRegistry.hasComponent<components::Jump>(entityID)){
-                newVelocity.y = updateJump(entityID, playerInput, velocity);
+                velocity->value.y = updateJump(entityID, playerInput, velocity);
             }
         }
+        updateState(entityID, velocity);
     }
 
-    if(velocity->value != newVelocity){
-        velocity->value = newVelocity;
+    if(velocity->value != oldValue){
         LOG_DEBUG_COMPONENT_CHANGED(name, velocity->name, velocity, entityID);
     }
 }
 
-float MovementSystem::updateHorizontalMovement(components::PlayerInput *playerInput, components::HorizontalMovement *horizontalMovement){
+float MovementSystem::updateHorizontalMovement(EntityID entityID, components::PlayerInput *playerInput, components::HorizontalMovement *horizontalMovement){
+    float newMovement = 0.f;
     if(playerInput->left){
+        newMovement = -horizontalMovement->speed;
         playerInput->left = false;
-        return -1.f * horizontalMovement->speed;
-    }
-
-    if (playerInput->right){
+    } 
+    else if (playerInput->right){
+        newMovement = horizontalMovement->speed;
         playerInput->right = false;
-        return 1.f * horizontalMovement->speed;
     }
 
-    return 0.f;
+    return newMovement;
+}
+
+void MovementSystem::updateState(EntityID entityID, components::Velocity *velocity){
+    if (mRegistry.hasComponent<components::State>(entityID)) {
+        auto state = mRegistry.getComponent<components::State>(entityID);
+        auto oldState = state->currentState;
+        auto oldDirection = state->direction;
+
+        if(velocity->value.y == 0 && velocity->value.x != 0){
+            state->currentState = components::State::CurrentState::WALK;
+        }
+        else if(velocity->value.y > 0){
+            state->currentState = components::State::CurrentState::FALL;
+        }
+        else if(velocity->value.y < 0){
+            state->currentState = components::State::CurrentState::JUMP;
+        }
+        else{
+            state->currentState = components::State::CurrentState::IDLE;
+        }
+
+        if(velocity->value.x < 0){
+            state->direction = components::State::Direction::LEFT;
+        }
+        else if(velocity->value.x > 0){
+            state->direction = components::State::Direction::RIGHT;
+        }
+        
+        if(state->currentState != oldState || state->direction != oldDirection){
+            LOG_DEBUG_COMPONENT_CHANGED(name, state->name, state, entityID);
+        }
+    }
 }
 
 float MovementSystem::updateJump(EntityID entityID, components::PlayerInput *playerInput, components::Velocity *velocity){
     auto jump = mRegistry.getComponent<components::Jump>(entityID);
+    float newVelocityY = velocity->value.y;
 
     if(playerInput->jump){
-        if(!mRegistry.hasComponent<components::Grounded>(entityID) || mRegistry.getComponent<components::Grounded>(entityID)->isGrounded){
-            return -jump->strength;
+        if(mRegistry.hasComponent<components::Grounded>(entityID) && mRegistry.getComponent<components::Grounded>(entityID)->isGrounded){
+            newVelocityY = -jump->strength;
         }
         playerInput->jump = false; 
     }
-    return velocity->value.y;
+
+    return newVelocityY;
 }
 
 const std::string MovementSystem::name = "MovementSystem";